% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconSeq.R
\name{deconSeq}
\alias{deconSeq}
\title{deconSeq}
\usage{
deconSeq(
  bulks,
  signature,
  weight = TRUE,
  intercept = TRUE,
  outdir = "./",
  prefix = "LinDeconSeq",
  verbose = TRUE
)
}
\arguments{
\item{outdir}{Output directory, default: current work direcory.}

\item{verbose}{logical, to print the detailed information for each iteration.}

\item{bulk}{bulk gene expression (gene x subject x measure).}

\item{pures}{cell type-specific gene expression matrice of cell types.}
}
\value{
Fractions of bulk sample
}
\description{
Cell type deoovolution for bulk samples based on weighted RLM method.
}
