% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{optimizeSignatures}
\alias{optimizeSignatures}
\title{Optimize the number of cell type-specific genes of each phnotype class to ensure the stability of the signature matrices.}
\usage{
optimizeSignatures(X, cell.names, min.group = 50, max.group = 200)
}
\arguments{
\item{X}{A list of cell markers that generated by the above step (assignedCellMarkers).}

\item{cell.names}{Target cell names of each subset.}

\item{min.group}{Minimum number of each phnotype class, default: 50.}

\item{max.group}{Maximum number of each phnotype class, default: 200.}
}
\value{
Derived signature matrices.
}
\description{
Optimize the number of cell type-specific genes of each phnotype class to ensure the stability of the signature matrices.
}
