% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconSeq.R
\name{deconSeq}
\alias{deconSeq}
\title{Deconvolution using w-RLM method.}
\usage{
deconSeq(
  bulks,
  signature,
  weight = TRUE,
  intercept = TRUE,
  scale = FALSE,
  QN = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bulks}{bulk gene expression (row = Genes, column = Samples).}

\item{signature}{Signature matrix, i.e. cell type-specific gene expression matrice of cell types.}

\item{weight}{Weight gene or not, default: TRUE.}

\item{intercept}{Add intercept when using robust linear regression, default: TRUE.}

\item{scale}{Scale bulk gene expression data or not, default: FALSE.}

\item{QN}{Quantile normalize bulk expression profile or not, default: FALSE}

\item{verbose}{logical, to print the detailed information for each iteration.}
}
\value{
A data.frame of fractions for bulk samples.
}
\description{
Cell type deoovolution for bulk samples based on weighted RLM method.
}
