% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{findMarkers}
\alias{findMarkers}
\title{Derive signature matrices.}
\usage{
findMarkers(refs, phes, min.group = 50, max.group = 200,
  norm.method = "TPM", q.cut = 0.01, p.cut = 0.05,
  data.type = "RNASeq", verbose = TRUE)
}
\arguments{
\item{refs}{Gene expression profile of each cell type, which rows represent genes and columns represent pure samples.}

\item{phes}{Phenotype classes. value 1 = indicate membership of the reference sample,
value 2 = indicates the class that the sample will be compared against.}

\item{min.group}{Minimum number of each phnotype class, default: 50.}

\item{max.group}{Maximum number of each phnotype class, default: 200.}

\item{norm.method}{Select method to normalize reference profile [QN, TPM, CPM, DESeq2 or NONE], default: TPM (Transcript per million).}

\item{data.type}{Data type, RNASeq, MA or PEAK, default: RNASeq.}

\item{verbose}{logical, to print the detailed information.}
}
\value{
list
}
\description{
Derive signature marker genes by a series steps, including normalizing, filtered low confidence genes and DEG analysis.
}
